Beginning of a rust version of rules
I'm thinking about being able to create a ruleset for the game inside
of rust. So the classes and everything will be able to be describable
and referencable inside of a rust app. For now I'm thinking of using
yaml to load up most of the data, but possibly allow for JSON or
something else in the future.

Eventually I might allow people to override some or all of the rules
in the ruleset. That's future though.